README.md
**B22-S2-IN1621 - Web Technologies (IT and AI) Module**

This repository contains materials and resources for the B22-S2-IN1621 Web Technologies (IT and AI) module, focusing on creating websites. Dive into web development fundamentals, explore the intersection of technology and artificial intelligence, and harness the power of modern tools and techniques to craft engaging and dynamic web experiences. Whether you're a beginner or seasoned developer, this module provides a comprehensive journey through the landscape of web development, equipping you with the skills and knowledge to bring your ideas to life on the web.